reorganize tests as described in comment of issue #12
