+++ test/test_fanuc.cpp
@@ -17,71 +17,8 @@ const std::string ROOT_LINK_PARAM = "root_link";
 // Robot description almost always called "robot_description" and therefore hardcoded below
 const std::string ROBOT_DESCRIPTION = "robot_description";
 
-/** \Brief Test ik and fk only using the plugin
- *
- * As the ik and fk are both calculated using the opw_kinematics packages,
- * this tests only check if passing values from the plugin to opw_kinematics works,
- * and if the ik and fk calculations inside opw_kinematics are consistent for this robot.
- */
-class TestPluginFanuc : public ::testing::Test
-{
-protected:
-  void SetUp() override
-  {
-    ros::NodeHandle pnh("~");
-    if (pnh.getParam(GROUP_PARAM, group_name_) && pnh.getParam(ROOT_LINK_PARAM, root_link_) &&
-        pnh.getParam(TIP_LINK_PARAM, tip_link_))
-    {
-      // the last parameter specifies "search_discretization", which is not used by the opw plugin
-      plugin_.initialize(ROBOT_DESCRIPTION, group_name_, root_link_, tip_link_, 0.1);
-    }
-    else
-    {
-      ROS_ERROR_STREAM("Failed to load parameters necessary to load plugin.");
-    }
-  }
-  void TearDown() override
-  {
-  }
-
-  moveit_opw_kinematics_plugin::MoveItOPWKinematicsPlugin plugin_;
-  std::string root_link_;
-  std::string tip_link_;
-  std::string group_name_;
-  std::string robot_description_name_;
-};
-
-TEST_F(TestPluginFanuc, InitOk)
-{
-  ASSERT_EQ(plugin_.getGroupName(), group_name_);
-}
-
-TEST_F(TestPluginFanuc, CompareIKAndFK)
-{
-  std::vector<std::string> link_names;
-  const std::vector<double> joint_angles = { 0, 0.1, 0.2, 0.3, 0.4, 0.5 };
-  std::vector<geometry_msgs::Pose> poses_out;
-
-  // find reachable pose
-  plugin_.getPositionFK(plugin_.getLinkNames(), joint_angles, poses_out);
-
-  // calculate all ik solutions for this pose
-  const std::vector<geometry_msgs::Pose> poses_in = { poses_out[0] };
-  std::vector<std::vector<double> > solutions;
-  kinematics::KinematicsResult result;
-  bool res = plugin_.getPositionIK(poses_in, joint_angles, solutions, result);
-  EXPECT_TRUE(res);
-
-  // check if fk for all this solutions gives the same pose
-  Eigen::Affine3d actual, desired;
-  tf::poseMsgToEigen(poses_out[0], desired);
-  for (auto js : solutions)
-  {
-    plugin_.getPositionFK(plugin_.getLinkNames(), js, poses_out);
-    tf::poseMsgToEigen(poses_out[0], actual);
-    moveit_opw_kinematics_plugin::testing::comparePoses(actual, desired);
-  }
-}
+// unused dummy value for kinematics solver instance
+const double DEFAULT_SEARCH_DISCRETIZATION = 0.01f;
 
 /** \Brief Check ik of opw_kinematics against fk of MoveIt! kinematics.
  *
@@ -115,7 +52,6 @@ protected:
   std::string root_link_;
   std::string tip_link_;
   std::string group_name_;
-  std::string robot_description_name_;
 
   robot_model::RobotModelPtr robot_model_;
   robot_state::RobotStatePtr robot_state_;
@@ -147,6 +83,8 @@ TEST_F(TestKinematicsFanuc, TestAllIkSingleJointPose)
   kinematics::KinematicsResult result;
   kinematics::KinematicsQueryOptions options;
   const auto solver = joint_model_group_->getSolverInstance();
+
+  solver->initialize(ROBOT_DESCRIPTION, group_name_, root_link_, tip_link_, DEFAULT_SEARCH_DISCRETIZATION);
 
   bool success = solver->getPositionIK(fk_poses, joint_angles, solutions, result, options);
   ASSERT_TRUE(success);
