+++ test/test_kuka.rostest
@@ -1,11 +1,15 @@
+<?xml version="1.0"?>
 <launch>
 
-<rosparam command="load" file="$(find moveit_opw_kinematics_plugin)/test/kinematics.yaml" />
-<include file="$(find kuka_kr6r700_moveit_config)/launch/planning_context.launch">
+<include file="$(find moveit_resources)/fanuc_moveit_config/launch/planning_context.launch">
   <arg name="load_robot_description" value="true"/>
 </include>
 
-<test pkg="moveit_opw_kinematics_plugin" type="moveit_opw_kinematics_plugin_test_kuka" test-name="test_kuka">
-</test>
+<rosparam command="load" file="$(find moveit_opw_kinematics_plugin)/test/kinematics_kuka.yaml" />
+<param name="tip_link" value="tool0" />
+<param name="root_link" value="base_link" />
+<param name="group" value="manipulator" />
+
+<test pkg="moveit_opw_kinematics_plugin" type="moveit_opw_kinematics_plugin_test_kuka" test-name="test_plugin_kuka"/>
 
 </launch>
